S2a: migrate the explicit WAVE state (wave_p/wave_v/wave_source) + mean_wp to Q16.16
First sub-step of the S2 atmosphere-group fixed-point migration. The explicit
wave system (AtmosphereSolver::wave_substep) and the mean_wp global reduction go
to int32 Q16.16; atmosphere/wind/smoke stay float (S2b/S2c) behind FLOAT BRIDGES.
Determinism-critical: integer +/-/* are exact + order-free, so the synced wave
state is now bit-identical run-to-run (cross-GPU once S2b/S2c land).

Q-S2-2 (measured, tests/_s2a_wave_v_measure.py): peak |wave_v| ~2674 under a
maximal blast << 32768 -> wave_v keeps Q16.16 (NO Q24.8). But |c_sq*lap| ~1.3e5
exceeds 32768 before *dt -> the velocity kick carries c_sq*lap and -damping*wave_v
in int64 and applies *dt BEFORE the narrow (dt-before-narrow, the only safe order).

fixed_point.h: new deterministic integer mean reduction (mean_sum int64 order-free
+ mean_round round-half-away-from-zero, sign-symmetric, no <<16 pre-shift) for
mean_wp. Unit-proved in tests/_s2a_mean_reduction_check.cpp (gated by
tests/test_s2a_mean_reduction.py): order-independent under shuffle, matches a
reference, sign-symmetric.

wave_substep now integer: source feed (mul_q16 + integer min), the Laplacian
(gather-once int64 face flux, narrow once), the kick (int64 dt-before-narrow),
the pressure update (mul_q16), the per-cell absorb (scale_mag, signed shrink),
the wall/vacuum BCs (exact 0), and mean_wp (rounded integer mean). The sponge BC
wave decays go integer too (scale_mag / mul_q16).

SUBTLE (the DC-leak guard): the wave->atmosphere transfer deposit is done in
FLOAT at the bridge (dequantize the integer anomaly, multiply by real transfer*dt)
ON PURPOSE — a truncating integer mul_q16 deposit DC-leaks (every cell loses 1 LSB
toward -inf -> ~percent/tick sink into the conserved atmosphere). The rounded mean
is DC-free; the float deposit keeps it unbiased + matches the float build. Sealed-
box atmosphere drift: 5.39e-5 (int) vs 5.38e-5 (float) — identical. Becomes a
round-to-nearest/conservative integer deposit at S2c (flagged in-code).

FLOAT BRIDGES (collapse at S2c): the transfer->atmosphere deposit, the wind read
of wave_p (dequantized in diffuse_solve), the per-face permeability weight
(quantized per face), and the water head/ripple reads of wave_p (dequantized in
physics_engine step_water/step_tail; the water TU stays float).

Python: gamemap allocates wave_p/wave_v/wave_source as int32; new wave_fixed.py
quantize/dequantize helpers; field_edit "wave" dtype (wave_source edits authored
in real units, stored Q16.16); recorder dequantizes at its debug boundary;
bindings expose WAVE_FIXEDPOINT and take int32 for the wave arrays.

Gates: P1 bit-determinism green (field_ab_harness asserts the wave fields int32 +
bit-identical run-to-run, 196 nonzero wave_p cells); feel-regression vs the float
build wave_p rel 3.8e-3 / wave_v rel 4.3e-3 (shockwave propagates + reflects the
same to Q16.16 granularity); no blowup/NaN under a 480-tick maximal blast (raw
<< int32max); full suite 354 passed; both --auto (default + unhcr_vessel_2) exit 0.
Full atmosphere conservation P2 is NOT testable yet (atmosphere is a float bridge
until S2c) — noted.

+++ src/simulation/gamemap.py
@@ -84,9 +84,17 @@ class GameMap:
         self.is_vacuum    = np.zeros((h, w), dtype=bool)
         self.flammable    = np.zeros((h, w), dtype=bool)
         self.atmosphere   = np.ones((h, w), dtype=np.float32)
-        self.wave_p       = np.zeros((h, w), dtype=np.float32)
-        self.wave_v       = np.zeros((h, w), dtype=np.float32)
-        self.wave_source  = np.zeros((h, w), dtype=np.float32)
+        # S2a: the explicit WAVE state is int32 Q16.16 (scale 2^16, shared with
+        # water/heat) — integer transport is bit-identical cross-machine (the
+        # determinism the float path lacked). wave_p (acoustic anomaly, signed),
+        # wave_v (velocity, signed), wave_source (injected energy, >= 0). The
+        # Q-S2-2 measurement pinned wave_v to Q16.16 (peak ~2674 << 32768). Field
+        # edits author wave_source in real units and quantize at the boundary
+        # (field_edit.py "wave" dtype); the renderer/recorder dequantize. See
+        # simulation.wave_fixed for the metres<->Q16.16 helpers.
+        self.wave_p       = np.zeros((h, w), dtype=np.int32)
+        self.wave_v       = np.zeros((h, w), dtype=np.int32)
+        self.wave_source  = np.zeros((h, w), dtype=np.int32)
         self.wind_x       = np.zeros((h, w), dtype=np.float32)
         self.wind_y       = np.zeros((h, w), dtype=np.float32)
         # Multi-gas density fields (engine/05 §6.2, M1): a dense (N, h, w) float32
